# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/appointments/patient/{patientId}": {
      "get": {
        "operationId": "AppointmentsController_findByPatient",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "description": "Patient user ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns appointments for the patient",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get appointments by patient ID",
        "tags": [
          "appointments"
        ]
      }
    },
    "/appointments/therapist/{therapistId}": {
      "get": {
        "operationId": "AppointmentsController_findByTherapist",
        "parameters": [
          {
            "name": "therapistId",
            "required": true,
            "in": "path",
            "description": "Therapist user ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns appointments for the therapist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get appointments by therapist ID",
        "tags": [
          "appointments"
        ]
      }
    },
...
    "/patient-profiles": {
      "post": {
        "operationId": "PatientProfileController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Profile created successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a patient profile",
        "tags": [
          "Patient Profiles"
        ]
      },
      "get": {
        "operationId": "PatientProfileController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile retrieved"
          },
          "404": {
            "description": "Profile not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a single patient profile",
        "tags": [
          "Patient Profiles"
        ]
      }
    },
    "/patient-profiles/all": {
      "get": {
        "operationId": "PatientProfileController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of profiles"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all patient profiles",
        "tags": [
          "Patient Profiles"
        ]
      }
    },
    "/patient-profiles/{id}": {
      "patch": {
        "operationId": "PatientProfileController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePatientProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a patient profile",
        "tags": [
          "Patient Profiles"
        ]
      },
      "delete": {
        "operationId": "PatientProfileController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile deleted successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a patient profile",
        "tags": [
          "Patient Profiles"
        ]
      }
    },
    "/therapist-profiles": {
      "post": {
        "operationId": "TherapistProfileController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTherapistProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TherapistProfile"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a therapist profile",
        "tags": [
          "Therapist Profiles"
        ]
      },
      "get": {
        "operationId": "TherapistProfileController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile retrieved"
          },
          "404": {
            "description": "Profile not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a single therapist profile",
        "tags": [
          "Therapist Profiles"
        ]
      }
    },
    "/therapist-profiles/all": {
      "get": {
        "operationId": "TherapistProfileController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TherapistProfile"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all therapist profiles",
        "tags": [
          "Therapist Profiles"
        ]
      }
    },
    "/therapist-profiles/{id}": {
      "patch": {
        "operationId": "TherapistProfileController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTherapistProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a patient profile",
        "tags": [
          "Therapist Profiles"
        ]
      },
      "delete": {
        "operationId": "TherapistProfileController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile deleted successfully"
          },
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a patient profile",
        "tags": [
          "Therapist Profiles"
        ]
      }
    },
    "/payments/stripe/create-payment-intent": {
      "post": {
        "operationId": "PaymentController_createPaymentIntent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStripePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StripePaymentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a payment intent in stripe",
        "tags": [
          "Payments"
        ]
      }
    },
    "/payments/webhook": {
      "post": {
        "operationId": "PaymentController_handleWebhook",
        "parameters": [
          {
            "name": "stripe-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Buffer"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Payments"
        ]
      }
    }
  },
  "info": {
    "title": "Therapist Booking System API",
    "description": "API documentation for the therapist booking system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TherapyType": {
        "type": "object",
        "properties": {}
      },
      "PatientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email"
        ]
      },
      "TherapistDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 3
          },
          "firstName": {
            "type": "string",
            "example": "Jane"
          },
          "lastName": {
            "type": "string",
            "example": "Smith"
          },
          "email": {
            "type": "string",
            "example": "jane.smith@example.com"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email"
        ]
      },
      "AppointmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "duration": {
            "type": "number",
            "example": 60
          },
          "scheduledAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-04-01T14:00:00Z"
          },
          "isCompleted": {
            "type": "boolean",
            "example": false
          },
          "meetingUrl": {
            "type": "string",
            "example": "https://zoom.us/j/123456789"
          },
          "therapyType": {
            "$ref": "#/components/schemas/TherapyType"
          },
          "patients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatientDto"
            }
          },
          "therapists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TherapistDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-03-24T12:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-03-24T12:00:00Z"
          },
          "isConfirmed": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "id",
          "duration",
          "scheduledAt",
          "isCompleted",
          "therapyType",
          "patients",
          "therapists",
          "createdAt",
          "updatedAt",
          "isConfirmed"
        ]
      },
...
      "CreatePatientProfileDto": {
        "type": "object",
        "properties": {
          "dateOfBirth": {
            "format": "date-time",
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          }
        }
      },
      "UpdatePatientProfileDto": {
        "type": "object",
        "properties": {
          "dateOfBirth": {
            "format": "date-time",
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          }
        }
      },
      "CreateTherapistProfileDto": {
        "type": "object",
        "properties": {
          "specialization": {
            "type": "string"
          },
          "licenseNumber": {
            "type": "string"
          },
          "yearsOfExperience": {
            "type": "number"
          },
          "userId": {
            "type": "number"
          }
        },
        "required": [
          "userId"
        ]
      },
      "TherapistProfile": {
        "type": "object",
        "properties": {}
      },
      "UpdateTherapistProfileDto": {
        "type": "object",
        "properties": {
          "specialization": {
            "type": "string"
          },
          "licenseNumber": {
            "type": "string"
          },
          "yearsOfExperience": {
            "type": "number"
          }
        }
      },
      "CreateStripePaymentDto": {
        "type": "object",
        "properties": {
          "appointmentId": {
            "type": "number",
            "description": "ID of the appointment being paid for",
            "example": 42
          },
          "amount": {
            "type": "number",
            "description": "Payment amount",
            "example": 99.99
          }
        },
        "required": [
          "appointmentId",
          "amount"
        ]
      },
      "StripePaymentResponseDto": {
        "type": "object",
        "properties": {
          "clientSecret": {
            "type": "string",
            "description": "Client secret for confirming the payment on the client side",
            "example": "pi_3LNtTY2eZvKYlo2C0hVh6XJQ_secret_3g2vV6XJQhVh6XJQhVh6XJQh"
          },
          "paymentId": {
            "type": "number",
            "description": "ID of the created payment record",
            "example": 12345
          }
        },
        "required": [
          "clientSecret",
          "paymentId"
        ]
      },
      "Buffer": {
        "type": "object",
        "properties": {}
      }
    }
  }
}